+++ app/src/main/java/ee/cyber/cdoc2/server/app/conf/SessionTokenConf.java
@@ -1,0 +1,5 @@
+package ee.cyber.cdoc2.server.app.conf;
+
+public interface SessionTokenConf {
+    String getIssuer();
+}